[FIX/#46] DataClient.OnDataChangedListener 추가
