+++ stempo/src/main/java/com/kkkk/stempo/presentation/WatchActivity.kt
@@ -10,12 +10,18 @@ import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
 import androidx.core.splashscreen.SplashScreen.Companion.installSplashScreen
+import com.google.android.gms.wearable.DataClient
+import com.google.android.gms.wearable.DataEvent
+import com.google.android.gms.wearable.DataEventBuffer
+import com.google.android.gms.wearable.DataMapItem
+import com.google.android.gms.wearable.Wearable
 import com.kkkk.stempo.presentation.home.HomeScreen
 import com.kkkk.stempo.presentation.theme.StempoandroidTheme
 import dagger.hilt.android.AndroidEntryPoint
+import timber.log.Timber
 
 @AndroidEntryPoint
-class WatchActivity : ComponentActivity() {
+class WatchActivity : ComponentActivity(), DataClient.OnDataChangedListener {
     override fun onCreate(savedInstanceState: Bundle?) {
         installSplashScreen()
 
@@ -28,5 +34,39 @@ class WatchActivity : ComponentActivity() {
                 HomeScreen()
             }
         }
+    }
+
+    override fun onResume() {
+        super.onResume()
+        Timber.tag("okhttp").d("LISTENER : ADDED")
+        Wearable.getDataClient(this).addListener(this)
+    }
+
+    override fun onPause() {
+        super.onPause()
+        Wearable.getDataClient(this).removeListener(this)
+    }
+
+    override fun onDataChanged(dataEvents: DataEventBuffer) {
+        Timber.tag("okhttp").d("LISTENER : ON DATA CHANGED")
+
+        dataEvents.forEach { event ->
+            if (event.type == DataEvent.TYPE_CHANGED) {
+                event.dataItem.also { item ->
+                    if (item.uri.path?.compareTo(PATH_BPM) == 0) {
+                        DataMapItem.fromDataItem(item).dataMap.apply {
+                            val bpm = getInt(KEY_BPM)
+                            Timber.tag("okhttp").d("LISTENER : DATA RECEIVED : $bpm")
+                        }
+                    }
+                }
+            }
+        }
+    }
+
+    companion object {
+        const val KEY_BPM = "KEY_BPM"
+
+        const val PATH_BPM = "/bpm"
     }
 }
